WIP: Use strongly typed worker profiles
The worker profiles are actually KubeletConfiguration manifests, and
there's ready-made Go types for it. Seems feasible to use those, instead
of untyped maps and slices.

This will probably not pan out because the kubelet's scheme is not part
of the exported kubernetes packages, and the k8s.io/kubernetes package
willi probably not be imported into k0s.

The interesting package is:
    k8s.io/kubernetes/pkg/kubelet/apis/config/scheme

See:
    https://stackoverflow.com/a/56183199
    https://github.com/golang/go/issues/32776
    https://stackoverflow.com/a/69296723

The solution without this is broken, since it will just null out any
defaults (e.g. the syncFrequency).

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>