Fix endless recursion in CallParentPersistentPreRun
The cobra code calls PersistentPreRun of parents with the actual
command, not with the parent command. CallParentPersistentPreRun did the
same, but that led to endless recursion, since it would find the same
method over and over again, when walking up the command stack. Fix this
by swapping out the pre-run methods temporarily.

Also:
* Improve kubeconfig detection, so that kubectl doesn't error out if the
  kubeconfig is specified as CLI arg
* Improve error reporting of the kubectl inttest
* Exclude tests from k0s's build dependencies in Makefile
* foo bar

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>