Explicitly set the embedded k8s binary versions
The auto-detection adds a plus to the minor git version when it detects
an alpha/beta/rc release, because k0s adds its +k0s suffix to the
version number. The plus would indicate that it's not the exact tagged
version, but that the tree is between two tagged versions. This is not
true for k0s-built kubernetes binaries.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>