Add depguard linter and forbid usage of gopkg.in/yaml
Since 543433f87, k0s is consistently using sigs.k8s.io/yaml. Add a lint
so that the former YAML lib is no longer in use.

Change the last occurrence of it in a way that no longer reads direcly
from the Reader, but buffers the HTTP response in memory. This is
required since sigs.k8s.io/yaml doesn't offer any API to use the Reader
directly, as it has to do two passes on the input anyways.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>