Add tenv linter to promote the usage of t.Setenv in tests
Refactor the sole test that used os.Setenv in a way that doesn't rely on
Setenv at all anymore, along with its implementation.

New implementation of supervisor's envForComponent:
Pass in the environment slice as first function argument. This enables
an easy way to test that function without mocking the OS environment.
Build up the merged env in a map along with a single iteration over the
env instead of sorting the env and using a map just for collision
detection. Add some helpful comments, too. All in all, the function
implementation should be easier to understand now.

New implementation of the test:
Reorder the test environment logically and add comments for the purpose
of the values. Use a test table to run both the test for keeping and
dropping prefixes. Add comments to the expected outcomes, too. Execute
the tests for each permutation of the input test env, so that it's
guaranteed that the implementation is agnostic about the order of its
input. Remove some env vars from the test env that didn't cover any new
cases to reduce the number of iterations. Allow for parallel execution
of these tests.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>