Run unit tests also on Windows
K0s doesn't support compilation on Windows, but it allows to be run as a
worker node there. There's currently no tests on or for Windows
whatsoever. (Despite the fact that it compiles.) By running the unit
tests on Windows, it can at least be ensured that k0s works correctly in
principle, i.e. when it comes to file system access, error handling and
such things.

Add a new GitHub Actions job that runs the tests on Windows. Not all of
the tests can be run there, or would even be meaningful on Windows in
the first place. Skip those tests based on runtime.GOOS using
t.Skip*(...) when appropriate. Add a build tag to those tests that use
UNIXy APIs and hence wouldn't compile there. Some tests required small
customizations concerning behavior of the file system (access patterns,
permissions, ...). Add special cases there for Windows, too.

Change the Autopilot tests from using the host OS/arch as labels to a
hard-coded version that always assumes linux/amd64. That removes the
limitation that a new entry has to be added to all of them for each new
OS/arch combination. In the end, the actual label values don't matter in
the tests.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>