Concurrent logs collection after inttests
Collect the k0s logs concurrently from all the nodes during the test
cleanup. Use tar on the controller node to assemble cluster state, so
that the client side logic gets simpler. Move the actual log reading
into launchDelegate, and let the OpenRC delegate collect the stderr
logs, too.

This is mainly to speed up things during shutdown, so that it's more
likely that the log collection succeeds before the tests time out.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>