Fail inttests when the deadline is exceeded
Some tests may not mark themselves as failed when the test context's
deadline is exceeded. This can result in green builds, even if the test
has been aborted. Moreover, the cluster state wasn't collected in those
cases, since the check if the test actually failed was likely to happen
before the integration tests recorded the failure.

Record a separate test failure whenever the context timed out, so that
the test is stamped as failed in any case and the cluster state will be
recorded.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>