Add GitHub job to run unit tests on Windows
K0s doesn't support compilation on Windows, but it allows to be run as a
worker node there. There's currently no tests on or for Windows
whatsoever. (Despite the fact that it compiles.) By running the unit
tests on Windows, it can at least be ensured that k0s works correctly in
principle, i.e. when it comes to file system access, error handling and
such things.