Fix certificate.Manager.EnsureCertificate
The file parameters to the cfssl library accept special colon-separated
prefixes that indicate if the parameter should be treated as a file path
or ann environmnet variable name. k0s didn't add those prefixes. This is
not a problem as long as the file path doesn't contain any colons. This
is rarely the case for Unix-like platforms. On Windows, though, any
absolute path will instantly hit that bug, since drive letters end with
a colon.

Fix this by explicitly adding the file: prefix to all file paths passed
to cfssl.

That's the unit test error on Windows before this change:

--- FAIL: TestCLITestSuite (0.35s)
    --- FAIL: TestCLITestSuite/TestKubeConfigCreate (0.35s)
        kubeconfig_test.go:127:
                Error Trace:    C:\Users\twieczorek\Repos\k0s\cmd\kubeconfig\kubeconfig_test.go:127
                Error:          Received unexpected error:
                                unknown prefix: C
                Test:           TestCLITestSuite/TestKubeConfigCreate