Add retries to kuberouter hairpinning inttest
There were quite some CI failures for this test that seemed to indicate
some transient errors. Try to make it more robust by adding a retry
mechanism. Replace the SSH/CLI based pod/exec approach with a client-go
based one. That way, the errors are more distinguishable and a decision
if something is to be retried becomes easier.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>