Lower wait timeouts for sonobuoy
The wait parameter is in minutes, so waiting 20 hours for the tests to
finish seems a bit too much. Currently, the conformance tests take less
than two hours and the network conformance tests less than ten minutes.
So wait four hours for the full conformance test suite and thirty
minutes for the network conformance test suite.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>