Rename files for uname syscall and add riscv64 build tag
The riscv64 architecture uses unsigned data types for the uname
syscall, just as the arm architecture does. Rename the files
containing the type definitions to signed and unsigned and add the
appropriate build tags for the architectures.

This enables k0s to be compiled on RISC-V 64-bit, when not using Docker,
i.e. via `make GO=go EMBEDDED_BINS_BUILDMODE=none`.