Add dir.Exists as an error-aware alternative to dir.IsDirectory
The dir.IsDirectory function is potentially hiding errors from callers.
In some cases, it's good to have a definitive answer to the question
if a directory really exists or not, or if it couldn't be determined
because of some OS error. The new function dir.Exists allows for that
distinction.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>