Unmount directories consistently
Have a single unount loop when deleting the data dir, instead of
potentially unmounting the same stuff in the containers step as well.
Also, by checking that the mount points are actually subdirectories of
the data dir, this is a safer approach than simply doing a substring
match on e.g. kubelet/pods and so on. Compile the unmount loop only
on Linux, as listing the mount points via the Kubernetes mounter was
a no-op on Windows anyway, and so we can use the syscall to unmount
instead of shelling out.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>