Reduce number of variables in TestGetEnv
The cases for "k1" and "k4" were identical, i.e. they both defined only
a generic variable. The cases for "k2" and "k3" were also identical,
i.e. they both defined generic and component-specific variables. Replace
the four cases  "k1" to "k4" with the three cases "only_generic", "both"
and "only_foo". This reduces the number of variables that need to be
defined from 10 to 8 while increasing the variation of the test input.
Also rename the variable values in the test to reflect where they're
coming from.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>