Use a context to propagate errors from first start
Using a context instead of a channel to propagate errors removes the
necessity to reason about channel buffers and potential blocking. It
makes both the error value as well as the started condition potentially
retrievable more than once. Moreover, blocking on a context's channel
that's cancelled in a defer block is easier to reason about than
blocking on a channel that's being sent to from a few branches and never
gets closed.

Use a shadowed variable for the started channel in the goroutine, so
that it can safely be set to nil after it has been closed. That way, the
first start doesn't have to be counted as a restart. Plus, this allows
for the channel to be gc'd after it has done its duty, too.

Only store cancel and done in the struct on a successful start. If the
process didn't start successful, the goroutine would exit anyways and
there's nothing to stop. Leaving the struct's done and cancel fields nil
in that case reflects the reality better, and also makes a new call to
Supervise() actually work

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>