Add controller lease counter component after API server
The controller lease counter requires the API server to be up and
running in order to work properly. It cannot acquire or count leases
before the API server is up, nor can it release the lease after the API
server is shut down.

Move the counter component down in the controller start method, so
that it's started after the API server. Decouple the actual controller
count value from the counter component by putting it directly into the
start method. Let the counter component update that value, and let the
konnectivity components consume those updates.

Introduce internal/sync.Latest as a small abstraction for values that
may change over time and will be consumed by multiple observers.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>