Disallow node-specific settings in ClusterConfig
Add CEL validation rules to disallow all the bits and pieces of the
configuration which are not reconcilable at the moment.

Add the IPv6 service CIDR to the list of non-reconcilable settings, for
the same reasons as its sibling.

Also add an extra CEL rule to check for the right Calico mode when
dual-stack is enabled.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>