Store single node flag directly in CfgVars
Store the single node flag as a bool pointer, so that it can be
determined whether the flag was actually present or not, if that might
ever matter in the future.

Decide directly in the method which default storage to use instead of
using a public field. Also, remove the unused setter method.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>