Decouple job names from metrics jobs
Introduce the Scraper interface that is now being used by the metrics
scraper component to obtain raw metrics that shall be pushed to the
pushgateway. Store the jobs in a map rather than in a slice. The map
key is used as the job name, hence the job struct itself doesn't need to
know its name anymore.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>