Hide certificate path juggling in certificate manager
Instead of passing the actual CA file paths in a certificate request,
use the CA name instead. The manager will do the rest by constructing
the file paths based on its root directory and the CA name. Also, if no
CA name is specified, let the manager automatically select the "default"
CA, named "ca".

The `Certificate` struct now carries the paths to the certificate
and key files, instead of the certificate and key data itself. This
is generally more useful, as the data itself was only used in `k0s
kubeconfig create`, which is now relying on the Kubernetes kubeconfig
API to load the data from disk.

Harmonize the `EnsureCA` and `EnsureCertificate` methods, so that they
both return a `Certificate` struct with the appropriate paths. This can
now be used by callers, so they don't have to rely on their own file
path arithmetic.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>