Don't check for closed watch channels
The only reason these channels get closed is if the watcher itself gets
closed. This happens only when the method returns, which in turn only
happens when the context is done. In this case, the loop has already
exited without a select on a potentially closed channel. So the branches
that checked for closed channels were effectively unreachable during
runtime.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>