Reuse worker's certificate manager in embedding controller
In controller+worker scenarios, the embedding controller used its own
certificate manager for the k0s status socket functionality that detects
worker-to-API connectivity. This is somewhat broken in case of NLLB.

Introduce an EmbeddingController interface that can be passed when
starting workers. This interface can be used to establish some
interaction between the embedded worker and its embedding controller.
Use this interface to share the certificate manager managed by the
worker with the controller. This way, the k0s status uses the same setup
that the actual worker uses.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>