Add tests for ClusterConfigInitializer
The DefaultClusterConfig function did not set the namespace, which is
required so that the fake clients know which namespace it belongs to
when it is added as an initial object. Also, this function hard-coded
some values for which there are constants.

For reactors to be added, the actual fake types are now exposed in the
fake client factory.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>