Share node name detection between controller and worker
The controller uses the node name solely when counting the number of
active controllers. Albeit it's not super-important to match the actual
worker node name in any case, it is probably a good practice and avoids
weird edge cases.

Use the Kubernetes API machinery's NodeName newtype to "tag" the node
name in a type safe manner, i.e. so that it can't be mixed with other
arbitrary strings.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>