Pull up filtering for secondary IPs
This is to encode the logic at the appropriate level. Surprisingly, the
interfaceIPs function filters IPs. Performing this task at the level at
which it is relevant makes the code clearer and allows interfaceIPs to
be reused in other contexts.

Introduce the new IP interface with an isSecondary method. On Linux,
the method will return the secondary flag; on other platforms, it will
return an unsupported error. Use this function in FirstPublicAddress.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>