Use deep default stripping on the whole config spec
Use the reflective/recursive config stripping for the whole config spec,
which allows to strip deeply nested defaults as well. Inline StripConfig
into GetClusterWideConfig, as it's only used in that context. This
allows for narrowing down the test cases, as some of the spec fields are
nilled out inconditionally anyways.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>