Overhaul docs templating
Allows for deeper integration into the sources and build system,
obtaining values from a single source of truth. This should reduce the
number of search-and-replace actions that typically occur across the
codebase.

Introduce a custom MkDocs macro module to enable more advanced
templating in the documentation. Add the src_var and build_var macros to
make values stored as literal variables or constants in the sources or
variables declared in the build system accessible.

Add the custom ljust filter to maintain the correct table layout when
replacing values of varying lengths.

Update the Makefiles and GitHub workflows to support the custom
module. Replace the KUBERNETES_VERSION environment variable with
build_var('kubernetes_version') and forward the k0s version to
the documentation build. Use the k0s version directly instead
of constructing it from the Kubernetes version in each template
replacement. Use the new macros and filters to replace some previously
hardcoded values.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>