Replace restarts with firstStart
The only important information about the restart count is whether a
result needs to be sent to the started channel. Have another channel
variable in the loop that will be set to nil as soon as the process
started successfully for the first time. Replace the restart counter
with this. Simplify the log statement and use structured logging for
the PID.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>