Split up Linux/Darwin process handling
There is no proc filesystem on macOS. Split things up and leave a
minimal implementation in place to reflect this fact. This allows most
of the unit tests to run on macOS while making it clear that proc stuff
doesn't exist on macOS.

Also enable the bogus PID file test unconditionally. This is just about
the PID file parsing, which doesn't require platform-specific APIs.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>