Revamp Calico manifests for Windows
Add all the Windows images to the ClusterConfig, and propagate the
Windows pause image via worker profiles.

Use templating for calico windows manifests instead of hardcoded CIDRs etc.

Provide the windows-specific default worker profile name directly in
the Cobra command, so that it's available right from the beginning, and
will be listed correctly in the help output as well.

Co-Authored-By: Jussi Nummelin <jnummelin@mirantis.com>
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>