Always generate SBOM, but only sign it when releasing
This removes some additional workflows and workflow runs by simply
building the SBOM after a successful build. This takes only a couple of
seconds, since all the requirements are already there.

Signing, on the other hand, is only required for release artifacts.

This commit also includes a necessary fix for the cosign 3.x bump: It's
now required to pass the --use-signing-config=false flag.

Fixes: 7986f1cf3 ("Update dependency sigstore/cosign to v3")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>