Refactor Dummy leader elector
Hardcode the dummy leader elector to always lead. A non-leading leader
elector was only required during testing, when a simple function was
sufficient.

Use a factory function to create the dummy leader elector and make the
never channel a private field of it instead of using a global variable.
In k0s, there will only ever be one dummy leader elector, if any.
Sharing the channel seemed like a good idea, but global state always
comes back to bite you. When using testing/synctest, that channel was
never associated with a bubble. This made synctest unusable with the
dummy elector.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>